refactor: domain 폴더 추가
