+++ src/main/java/roomescape/domain/reservation/dto/ReservationResponse.java
@@ -1,8 +1,8 @@
-package roomescape.reservation.dto;
+package roomescape.domain.reservation.dto;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
-import roomescape.reservation.domain.Reservation;
+import roomescape.domain.reservation.domain.Reservation;
 
 public record ReservationResponse(Long id, String name, LocalDate date, String time) {
