+++ src/main/java/roomescape/domain/reservation/exception/ReservationNotFoundException.java
@@ -1,4 +1,4 @@
-package roomescape.reservation.exception;
+package roomescape.domain.reservation.exception;
 
 import roomescape.global.exception.RoomescapeNotFoundException;
 import roomescape.global.exception.code.ErrorStatus;
